refactor: Split screen into small composables
