+++ app/src/main/java/com/emenjivar/luminar/screen/camera/CameraScreen.kt
@@ -1,16 +1,22 @@
 package com.emenjivar.luminar.screen.camera
 
 import android.Manifest
-import android.content.Context
 import android.graphics.Bitmap
 import android.widget.ImageView
-import androidx.camera.core.Camera
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.ImageCapture
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.camera.view.PreviewView
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.animateScrollBy
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -19,27 +25,30 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.LifecycleOwner
 import com.emenjivar.luminar.R
 import com.emenjivar.luminar.ext.settingsIntent
 import com.emenjivar.luminar.ui.components.CustomDialog
 import com.emenjivar.luminar.ui.components.CustomDialogAction
+import com.emenjivar.luminar.ui.components.MorseText
+import com.emenjivar.luminar.ui.components.SwitchButton
 import com.emenjivar.luminar.ui.components.rememberCustomDialogController
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
 
 @Composable
 fun CameraScreen(
@@ -48,7 +57,7 @@ fun CameraScreen(
     CameraScreenContent(uiState = viewModel.state)
 }
 
-@OptIn(ExperimentalPermissionsApi::class)
+@OptIn(ExperimentalPermissionsApi::class, ExperimentalMaterial3Api::class)
 @androidx.annotation.OptIn(androidx.camera.core.ExperimentalGetImage::class)
 @Composable
 fun CameraScreenContent(
@@ -56,9 +65,9 @@ fun CameraScreenContent(
 ) {
     // Compose variables
     val context = LocalContext.current
-    val lifecycleOwner = LocalLifecycleOwner.current
 
     // Controllers
+    val verticalScroll = rememberScrollState()
     val dialogController = rememberCustomDialogController()
     val permissionState = rememberPermissionState(
         permission = Manifest.permission.CAMERA,
@@ -77,24 +86,28 @@ fun CameraScreenContent(
     val blobRadiusRange by uiState.blobRadiusRange.collectAsState()
 
     // Remembered values
+    val verticalJumpPx = with(LocalDensity.current) { verticalJump.toPx() }
+    val isDebugEnabled = remember { mutableStateOf(false) }
+    val shouldDisplaySettings = remember { mutableStateOf(false) }
+
+    var imageWithFilter by remember {
+        mutableStateOf<Bitmap?>(null)
+    }
+    val isFlashTurnOn = remember { mutableStateOf(false) }
+
     val previewView = remember {
         PreviewView(context).apply {
             this.scaleType = PreviewView.ScaleType.FILL_CENTER
         }
     }
-    val imageCapture = remember { ImageCapture.Builder().build() }
-    val cameraSelector = remember {
-        CameraSelector.Builder()
-            .requireLensFacing(CameraSelector.LENS_FACING_BACK)
-            .build()
+    val imageAnalysis = remember {
+        CustomImageAnalyzer(
+            onDrawImage = { isTurnOn, bitmap ->
+                isFlashTurnOn.value = isTurnOn
+                imageWithFilter = bitmap
+            }
+        )
     }
-    val executor = remember {
-        ContextCompat.getMainExecutor(context)
-    }
-    var imageWithFilter by remember {
-        mutableStateOf<Bitmap?>(null)
-    }
-    val isFlashTurnOn = remember { mutableStateOf(false) }
 
     // SideEffects
     LaunchedEffect(isFlashTurnOn) {
@@ -103,6 +116,11 @@ fun CameraScreenContent(
             .onEach { isTurnOn ->
                 uiState.addFlashState(isTurnOn)
             }.launchIn(this)
+    }
+
+    LaunchedEffect(messages) {
+        // Scroll to the bottom after receiving a new message
+        verticalScroll.animateScrollBy(verticalJumpPx * messages.size)
     }
 
     // TODO: this block could be moved to the viewModel
@@ -128,71 +146,108 @@ fun CameraScreenContent(
         }
     }
 
-    val imageAnalysis = remember {
-        ImageAnalysis.Builder()
-            .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
-            .build().apply {
-                setAnalyzer(executor, CustomImageAnalyzer(
-                    onDrawImage = { isTurnOn, bitmap ->
-                        isFlashTurnOn.value = isTurnOn
-                        imageWithFilter = bitmap
-                    }
-                ))
-            }
-    }
-    var camera by remember { mutableStateOf<Camera?>(null) }
-
-    // States
-    val preview = Preview.Builder().build()
-
-    // Launched
-    LaunchedEffect(Unit) {
-        permissionState.launchPermissionRequest()
-    }
-
-    LaunchedEffect(permissionState.status) {
-        if (permissionState.status.isGranted) {
-            camera = startCamera(
-                context = context,
-                lifecycleOwner = lifecycleOwner,
-                cameraSelector = cameraSelector,
-                preview = preview,
-                previewView = previewView,
-                imageCapture = imageCapture,
-                imageAnalysis = imageAnalysis
-            )
-        }
-    }
-
-    CameraScreenLayout(
-        morse = debugMorse,
-        messages = messages,
-        circularityRange = { circularityRange },
-        blobRadiusRange = { blobRadiusRange },
-        onSetCircularity = uiState.onSetCircularity,
-        onSetBlobRadius = uiState.onSetBlobRadius,
-        rawCameraPreview = { modifier ->
-            AndroidView(
-                modifier = modifier,
-                factory = { previewView }
-            )
-        },
-        filteredCameraPreview = { modifier ->
-            AndroidView(
-                modifier = modifier,
-                factory = { context ->
-                    ImageView(context).apply {
-                        scaleType = ImageView.ScaleType.FIT_CENTER
-                    }
-                },
-                update = { view ->
-                    imageWithFilter?.let { bitmap ->
-                        view.setImageBitmap(bitmap)
+    Scaffold(
+        contentColor = Color.Transparent,
+        topBar = {
+            TopAppBar(
+                title = {},
+                colors = TopAppBarDefaults.smallTopAppBarColors(
+                    containerColor = Color.Transparent
+                ),
+                actions = {
+                    Box(
+                        modifier = Modifier
+                            .clip(CircleShape)
+                            .background(Color.White.copy(alpha = 0.2f))
+                            .padding(8.dp)
+                            .clickable {
+                                shouldDisplaySettings.value = !shouldDisplaySettings.value
+                            },
+                    ) {
+                        Icon(
+                            modifier = Modifier
+                                .clip(CircleShape),
+                            painter = painterResource(
+                                id = if (shouldDisplaySettings.value) {
+                                    R.drawable.ic_close
+                                } else {
+                                    R.drawable.ic_settings
+                                }
+                            ),
+                            contentDescription = "Settings",
+                            tint = Color.White
+                        )
                     }
                 }
             )
         }
-    )
+    ) { paddingValues ->
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues.calculateBottomPadding())
+        ) {
+            DualCameraPreview(
+                isDebugEnabled = isDebugEnabled.value,
+                rawPreview = { modifier ->
+                    LiveCameraPreview(
+                        modifier = modifier,
+                        previewView = previewView,
+                        analyzer = imageAnalysis,
+                        permissionState = permissionState
+                    )
+                },
+                debugPreview = { modifier ->
+                    AndroidView(
+                        modifier = modifier,
+                        factory = { context ->
+                            ImageView(context).apply {
+                                scaleType = ImageView.ScaleType.FIT_CENTER
+                            }
+                        },
+                        update = { view ->
+                            imageWithFilter?.let { bitmap ->
+                                view.setImageBitmap(bitmap)
+                            }
+                        }
+                    )
+                }
+            )
+
+            if (shouldDisplaySettings.value) {
+                CameraSettings(
+                    modifier = Modifier.align(Alignment.BottomCenter),
+                    circularityRange = { circularityRange },
+                    blobRadiusRange = { blobRadiusRange },
+                    onSetCircularity = uiState.onSetCircularity,
+                    onSetBlobRadius = uiState.onSetBlobRadius
+                )
+            } else {
+                MessageHistory(
+                    modifier = Modifier
+                        .align(Alignment.BottomCenter),
+                    messages = messages,
+                    verticalScroll = verticalScroll
+                )
+                SwitchButton(
+                    modifier = Modifier
+                        .align(Alignment.BottomEnd),
+                    text = "Debug",
+                    isEnabled = isDebugEnabled.value,
+                    onEnable = { isEnabled ->
+                        isDebugEnabled.value = isEnabled
+                    }
+                )
+            }
+
+            if (debugMorse.isNotBlank()) {
+                MorseText(
+                    modifier = Modifier.align(Alignment.Center),
+                    text = debugMorse
+                )
+            }
+        }
+    }
 
     CustomDialog(
         controller = dialogController,
@@ -211,33 +266,4 @@ fun CameraScreenContent(
     )
 }
 
-private suspend fun startCamera(
-    context: Context,
-    lifecycleOwner: LifecycleOwner,
-    cameraSelector: CameraSelector,
-    preview: Preview,
-    previewView: PreviewView,
-    imageCapture: ImageCapture,
-    imageAnalysis: ImageAnalysis
-): Camera {
-    val cameraProvider = context.getCameraProvider()
-    cameraProvider.unbindAll()
-    val camera = cameraProvider.bindToLifecycle(
-        lifecycleOwner,
-        cameraSelector,
-        preview,
-        imageAnalysis,
-        imageCapture
-    )
-    preview.setSurfaceProvider(previewView.surfaceProvider)
-    return camera
-}
-
-private suspend fun Context.getCameraProvider(): ProcessCameraProvider =
-    suspendCoroutine { continuation ->
-        val cameraProviderFuture = ProcessCameraProvider.getInstance(this)
-        val provider = cameraProviderFuture.get()
-        cameraProviderFuture.addListener({
-            continuation.resume(provider)
-        }, ContextCompat.getMainExecutor(this))
-    }
+private val verticalJump = 20.dp
